Added new `const` values for non-option arguments and "--".
Added check to determine whether 0x01 must be returned, depending on whether or not `ShortOpts[0] == '-'`.

Added check to determine if parsing should stop, depending on whether or not `ShortOpts[0] == '+'`.

If a double-dash (`--`) is encountered by `GetNextOpt()` and `DoubleDashStopsParsing` is enabled, then `-1` is returned.

If `IgnoreInvalidOptions` is `true` and `MustReturnChar1()` returns `true`, then `NonOptChar` is returned.